server: move hover to destinated directory
