+++ language-server/src/features/hover/hover.ts
@@ -1,20 +1,20 @@
 import { inject, injectable } from 'tsyringe'
 import { Connection, MarkupContent } from 'vscode-languageserver'
-import { Provider } from './provider'
+import { Provider } from '../provider'
 import * as winston from 'winston'
 import * as ls from 'vscode-languageserver'
-import { RangeConverter } from '../utils/rangeConverter'
+import { RangeConverter } from '../../utils/rangeConverter'
 import { URI } from 'vscode-uri'
-import { LoadFolder } from '../mod/loadfolders'
-import { ProjectManager } from '../projectManager'
-import { Project } from '../project'
+import { LoadFolder } from '../../mod/loadfolders'
+import { ProjectManager } from '../../projectManager'
+import { Project } from '../../project'
 import { Element } from '@rwxml/analyzer'
 import { AsEnumerable } from 'linq-es2015'
-import { TextDocumentManager } from '../textDocumentManager'
-import { Definition } from './definition'
-import { FileStore } from '../fileStore'
-import { DependencyFile } from '../fs'
-import { LogToken } from '../log'
+import { TextDocumentManager } from '../../textDocumentManager'
+import { Definition } from '../definition'
+import { FileStore } from '../../fileStore'
+import { DependencyFile } from '../../fs'
+import { LogToken } from '../../log'
 // how to use 'prettydiff' (it is quite different to use than other standard libs)
 // https://github.com/prettydiff/prettydiff/issues/176
 // https://github.com/sprity/sprity/blob/master/lib/style.js#L38-L53
